more journey test cleanup
