+++ package.json
@@ -21,7 +21,7 @@
     "test:journey:k3d": "k3d cluster delete pepr-dev && k3d cluster create pepr-dev --k3s-arg '--debug@server:0'",
     "test:journey:build": "npm run build && npm pack",
     "test:journey:image": "docker buildx build --tag pepr:dev . && k3d image import pepr:dev -c pepr-dev",
-    "test:journey:run": "ava journey/*.test.ts --sequential --timeout=2m",
+    "test:journey:run": "ava journey/entrypoint.test.ts --timeout=2m",
     "format:check": "eslint src && prettier src --check",
     "format:fix": "eslint src --fix && prettier src --write"
   },
